requirements.txt
flask
pillow
numpy
matplotlib
scikit-learn
jinja2
